# back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API DoReMi - Plateforme scolaire numérique",
        "description": "Bienvenue dans la documentation de l'API **DoReMi**, une plateforme complète de gestion scolaire hybride.\\n\\n\n        ## Modules couverts :\\n\n        - Gestion des écoles et utilisateurs (élèves, enseignants, parents)\\n\n        - Cours en ligne et en présentiel\\n\n        - Suivi des notes et bulletins\\n\n        - Gestion des documents et devoirs\\n\n        - Actualités, emploi du temps, messagerie\\n\n        - Gestion des stages : demandes & offres\\n\\n\n        Cette documentation est à jour pour l'environnement de développement.",
        "termsOfService": "https://doremi-app.test/conditions",
        "contact": {
            "name": "Support Technique DoReMi",
            "email": "support@doremi-app.test"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local de développement"
        }
    ],
    "paths": {
...
        "/api/search": {
            "get": {
                "tags": [
                    "Recherche"
                ],
                "summary": "Recherche globale sur toutes les ressources",
                "operationId": "b40b0ddf237f237094f0a5f59191bd28",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Terme de recherche",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type de ressource (courses, documents, internships, news, videos)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre de résultats par type",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Résultats de recherche"
                    }
                }
            }
        },
...
        "/api/support/tickets/{userId}": {
            "get": {
                "tags": [
                    "Support"
                ],
                "summary": "Obtenir les tickets de support d'un utilisateur",
                "operationId": "b81d73ede9868439ce82bbe369511c46",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrer par statut",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "open",
                                "in_progress",
                                "resolved",
                                "closed"
                            ]
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "description": "Filtrer par priorité",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "low",
                                "medium",
                                "high",
                                "urgent"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tickets de support récupérés avec succès"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/support/tickets": {
            "post": {
                "tags": [
                    "Support"
                ],
                "summary": "Créer un nouveau ticket de support",
                "operationId": "455f413d21a26ea23e9089fe5e69a658",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "subject": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "category": {
                                        "type": "string",
                                        "enum": [
                                            "technical",
                                            "billing",
                                            "course",
                                            "general"
                                        ]
                                    },
                                    "priority": {
                                        "type": "string",
                                        "enum": [
                                            "low",
                                            "medium",
                                            "high",
                                            "urgent"
                                        ]
                                    },
                                    "attachments": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ticket de support créé avec succès"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/support/tickets/{id}": {
            "get": {
                "tags": [
                    "Support"
                ],
                "summary": "Obtenir un ticket de support spécifique",
                "operationId": "a52cbb6124fa23d998763fcfdf4f7e8b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du ticket",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket de support récupéré avec succès"
                    },
                    "404": {
                        "description": "Ticket non trouvé"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Support"
                ],
                "summary": "Mettre à jour un ticket de support",
                "operationId": "0c9bea26416fa77dd831674627407da4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du ticket",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string"
                                    },
                                    "priority": {
                                        "type": "string"
                                    },
                                    "assigned_to": {
                                        "type": "integer"
                                    },
                                    "admin_notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ticket de support mis à jour avec succès"
                    },
                    "404": {
                        "description": "Ticket non trouvé"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Support"
                ],
                "summary": "Supprimer un ticket de support",
                "operationId": "2e70fb8d0e14e10ad52104eeaaf673bb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du ticket",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket de support supprimé avec succès"
                    },
                    "404": {
                        "description": "Ticket non trouvé"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/support/faq": {
            "get": {
                "tags": [
                    "Support"
                ],
                "summary": "Obtenir la FAQ",
                "operationId": "77e8c3c33a012673d5149b0c621fbeff",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filtrer par catégorie",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Rechercher dans la FAQ",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "FAQ récupérée avec succès"
                    }
                }
            },
            "post": {
                "tags": [
                    "Support"
                ],
                "summary": "Créer une nouvelle entrée FAQ",
                "operationId": "d4d599bdbfeebfbc0887140711d2cd60",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "question": {
                                        "type": "string"
                                    },
                                    "answer": {
                                        "type": "string"
                                    },
                                    "category": {
                                        "type": "string"
                                    },
                                    "order": {
                                        "type": "integer"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "FAQ créée avec succès"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/support/faq/{id}": {
            "put": {
                "tags": [
                    "Support"
                ],
                "summary": "Mettre à jour une entrée FAQ",
                "operationId": "f40896372cdfa42f7fcb514660cb294f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la FAQ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "question": {
                                        "type": "string"
                                    },
                                    "answer": {
                                        "type": "string"
                                    },
                                    "category": {
                                        "type": "string"
                                    },
                                    "order": {
                                        "type": "integer"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "FAQ mise à jour avec succès"
                    },
                    "404": {
                        "description": "FAQ non trouvée"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Support"
                ],
                "summary": "Supprimer une entrée FAQ",
                "operationId": "5a4f0bf4871b9a8853020ebf753bd0b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la FAQ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "FAQ supprimée avec succès"
                    },
                    "404": {
                        "description": "FAQ non trouvée"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}